Moved ProjectsProvider and ProjectsProviderAPI to the projects module
