+++ collect_app/src/test/java/org/odk/collect/android/provider/ProjectsProviderTest.kt
@@ -11,9 +11,10 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.odk.collect.android.injection.DaggerUtils
-import org.odk.collect.android.provider.ProjectsProvider.Companion.PROJECT_NAME
-import org.odk.collect.android.provider.ProjectsProvider.Companion.PROJECT_UUID
 import org.odk.collect.projects.Project
+import org.odk.collect.projects.providers.ProjectsProvider.Companion.PROJECT_NAME
+import org.odk.collect.projects.providers.ProjectsProvider.Companion.PROJECT_UUID
+import org.odk.collect.projects.providers.ProjectsProviderAPI
 
 @RunWith(AndroidJUnit4::class)
 class ProjectsProviderTest {
