+++ projects/src/main/java/org/odk/collect/projects/DaggerSetup.kt
@@ -3,6 +3,7 @@ package org.odk.collect.projects
 import dagger.Component
 import dagger.Module
 import dagger.Provides
+import org.odk.collect.projects.providers.ProjectsProvider
 import org.odk.collect.shared.UUIDGenerator
 import javax.inject.Singleton
 
@@ -15,6 +16,8 @@ interface ProjectsDependencyComponentProvider {
 interface ProjectsDependencyComponent {
 
     fun inject(addProjectDialog: AddProjectDialog)
+
+    fun inject(projectsProvider: ProjectsProvider)
 }
 
 @Module
